+++ lib/common/utils/pubspec/pubspec_lock.dart
@@ -1,38 +1,4 @@
-import 'dart:io';
+import 'package_version.dart';
 
-import 'package:path/path.dart';
-import 'package:yaml/yaml.dart';
-
-import '../../../core/internationalization.dart';
-import '../../../core/locales.g.dart';
-import '../../../functions/version/check_dev_version.dart';
-import '../logger/log_utils.dart';
-
-class PubspecLock {
-  static Future<String?> getVersionCli({bool disableLog = false}) async {
-    try {
-      // TODO: 本地打包获取版本号
-      var scriptFile = Platform.script.toFilePath();
-      var pathToPubLock = join(dirname(scriptFile), '../pubspec.lock');
-      final file = File(pathToPubLock);
-      var text = loadYaml(await file.readAsString());
-      if (text['packages']['sm_get_cli'] == null) {
-        if (isDevVersion()) {
-          if (!disableLog) {
-            LogService.info('Development version');
-          }
-        }
-        return null;
-      }
-      var version = text['packages']['sm_get_cli']['version'].toString();
-      return version;
-    } on Exception catch (_) {
-      if (!disableLog) {
-        LogService.error(
-          Translation(LocaleKeys.error_cli_version_not_found).tr,
-        );
-      }
-      return null;
-    }
-  }
-}
+@Deprecated('Use PackageVersion from package_version.dart instead.')
+typedef PubspecLock = PackageVersion;